pyproject.toml
[project]
name = "chainlit-litellm-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.2.1",
    "litellm>=1.63.3",
    "python-dotenv>=1.0.1",
]
